backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
sqlalchemy>=2.0.25
aiosqlite>=0.19.0  # Async SQLite driver
asyncpg>=0.29.0  # Async PostgreSQL driver for Neon

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# Environment variables
python-dotenv>=1.0.0

# Phase III - AI Chat
groq>=0.4.1        # Groq API SDK
mcp>=0.9.0         # Official MCP SDK

# Type checking and linting (optional but recommended)
mypy>=1.8.0
types-pyyaml>=6.0.0
types-pyjwt>=1.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
